Manage goals add gear, green cards, and direct link to create goals
Changes:

  - Add grey gear that floats to the right of app cards on manage goals.
  - Turn cards green if a goal has been sent.
  - On click, send apps without any goals set to the create a goal page.

+++ app/src/main/java/com/blue_stingray/healthy_life_app/model/Goal.java
@@ -6,18 +6,18 @@ public class Goal {
     private String application_id;
     private String packageName;
 
-    private transient int timeLimit;
-    private transient int limitDay;
+    private transient float timeLimit;
+    private transient float limitDay;
 
     public String getPackageName() {
         return packageName;
     }
 
-    public int getTimeLimit() {
+    public float getTimeLimit() {
         return timeLimit;
     }
 
-    public int getLimitDay() {
+    public float getLimitDay() {
         return limitDay;
     }
 
@@ -25,11 +25,11 @@ public class Goal {
         this.packageName = packageName;
     }
 
-    public void setTimeLimit(int timeLimit) {
+    public void setTimeLimit(float timeLimit) {
         this.timeLimit = timeLimit;
     }
 
-    public void setLimitDay(int limitDay) {
+    public void setLimitDay(float limitDay) {
         this.limitDay = limitDay;
     }
 }
